refactor: :art: move legend component
